move attachment endpoints that do not require issue id
