+++ VisconSupportAPI/Controllers/IssueController.cs
@@ -251,98 +251,6 @@ public class IssueController: BaseController
             ID = newAttachment.Id
         });
     }
-    
-    [HttpGet("{issueId:int}/attachments/{attachmentId:int}")]
-    // [Authorize]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-    [ProducesResponseType(StatusCodes.Status403Forbidden)]
-    [ProducesResponseType(StatusCodes.Status403Forbidden)]
-    public ActionResult<List<Attachment>> GetAttachment(int issueId, int attachmentId)
-    {
-        // User? user = GetUserFromClaims();
-        // if (user == null)
-        // {
-        //     return Unauthorized();
-        // }
-
-        Issue? selectedIssue = Context.Issues.FirstOrDefault(i => i.Id == issueId);
-
-        if (selectedIssue == null)
-        {
-            return NotFound();
-        }
-
-        // if (user.Type is not (AccountType.Admin or AccountType.Helpdesk) && selectedIssue.UserId != user.Id)
-        // {
-        //     return Forbid();
-        // }
-
-        Attachment? attachment = Context.Attachments.FirstOrDefault(a => a.Id == attachmentId);
-
-        if (attachment == null)
-        {
-            return NotFound();
-        }
-
-        string? ext = Utils.GetFileExtension(attachment.MimeType);
-
-        if (ext == null)
-        {
-            return StatusCode(500, new
-            {
-                Message = "Invalid file mimetype"
-            });
-        }
-
-        string path = Path.Join(Configuration["DataPath"], $"{attachment.Id}{ext}");
-
-        FileStream stream = new FileStream(path, FileMode.Open, FileAccess.Read);
-        return File(stream, attachment.MimeType);
-    }
-    
-    [HttpGet("attachments/{attachmentId:int}/authenticate")]
-    [Authorize]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-    [ProducesResponseType(StatusCodes.Status403Forbidden)]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    public ActionResult AuthenticateAttachment(int attachmentId, string mimeType)
-    {
-        User? user = GetUserFromClaims();
-        
-        if (user == null)
-        {
-            return Unauthorized();
-        }
-        
-        Attachment? attachment = Context.Attachments.FirstOrDefault(a => a.Id == attachmentId);
-
-        if (attachment == null)
-        {
-            return NotFound();
-        }
-
-        if (attachment.MimeType != mimeType)
-        {
-            return BadRequest();
-        }
-
-        Issue? selectedIssue = Context.Issues.FirstOrDefault(i => i.Id == attachment.IssueId);
-
-        if (selectedIssue == null)
-        {
-            return NotFound();
-        }
-
-        if (user.Type is not (AccountType.Admin or AccountType.Helpdesk) && selectedIssue.UserId != user.Id)
-        {
-            return Forbid();
-        }
-
-        return Ok();
-    }
 }
 
 public class NewAttachment
